Squashed commit of the following:
commit a790db3aec09be45fb257c9a0ff3a8d1feba486a
Author: AJ Slater <aj@slater.net>
Date:   Tue May 9 17:41:04 2023 -0700

    copse -> grove

commit f399ec0fd1dc83b2bb42636305c94726d5685719
Author: AJ Slater <aj@slater.net>
Date:   Tue May 9 12:58:42 2023 -0700

    resolve get_dir() paths

commit 4c2c1667a927e9be54ce8d55e2ee3ab03c3f5aa9
Author: AJ Slater <aj@slater.net>
Date:   Mon May 8 18:06:03 2023 -0700

    resolve paths in gettiem and settitem for copse. also everything int _to_absolute_paths

commit 4a5fcd510537e2d51bc42c2dff4d5b3a8ff93a0a
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 19:18:20 2023 -0700

    fix elif error

commit 80b5fa0221e0d5ea25fbac1679e9f287a0d571c7
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 19:02:01 2023 -0700

    don't save sets just lists

commit 8c555ea3aae2d3e49383241520bee43153bae5dd
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 18:49:11 2023 -0700

    use the normalizer on the treestamps config values

commit ec97bcf0b4e1024de4d2f9a12dbac07e38551028
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 18:37:34 2023 -0700

    consolidate some dump code

commit 5c0add147e53312f8ff39b1acf76e591939d7080
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 18:37:24 2023 -0700

    resolve root dir

commit 59e4192b461032cd962196243f7ef4eb04447223
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 18:37:04 2023 -0700

    update deps

commit 3872adb7c124cfb72dd664858028cf4a12814538
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 18:36:58 2023 -0700

    install package not file now

commit cca0c53c5e227ce8d74caf4aac984ac9b724d0e8
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 15:30:27 2023 -0700

    update docs

commit 7eab76d344c651f66a68a9248aa99073749b67e3
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 15:25:30 2023 -0700

    put ignore and symlinks pref in config file. add check_config option

commit 14627893d3b498b6e09faa9eaa89db9e1b0bca94
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 15:04:37 2023 -0700

    big config refactor

commit 2fdadd9a685f1c4314ebf1e4902d5318bc313b47
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 13:16:09 2023 -0700

    update readme

commit 9cd2f8c444bf46ee66ecb96be5577b6baad0531f
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 13:09:48 2023 -0700

    reorg tree into its own dir. move write methods into own file

commit 0c96656aef085b0c921feaba284bd47cc02ed51f
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 12:58:03 2023 -0700

    reorg methods a bit

commit b7f502e82d493a1125dc33dedbd9384044a5db01
Author: AJ Slater <aj@slater.net>
Date:   Fri May 5 12:47:46 2023 -0700

    copse and new config basically working

commit da79cad4479fb1ed3966bfc483b9694fe6ab3521
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 17:25:25 2023 -0700

    make consume_all_child_timestamps private

commit 546b4b986729a449bc8a49a6cd929e4b081e3ed4
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 17:23:21 2023 -0700

    make prune_dict() private

commit fdd737bb31e5ad6dda34c328718cf915dfd187df
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 17:14:05 2023 -0700

    fix README for last change

commit 4f94528a3a7b78492ec4030c9562c0cec9fed112
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 17:13:32 2023 -0700

    self.dir (self.dir_path) -> self.root_dir

commit d63fbaa791252c52d477e2076105177f5084ece7
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 17:08:56 2023 -0700

    dirpath() -> get_dir()

commit 5035c5a2c7a2fc7ee010f9dcf43367de99721dbc
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 10:46:14 2023 -0700

    make treestamps much more modular

commit c523748afb1b2b2b75e4be621722463d0323fd57
Author: AJ Slater <aj@slater.net>
Date:   Thu May 4 10:45:55 2023 -0700

    fix version for no frontend
